+++ docs/CHANGELOG.md
@@ -11,7 +11,14 @@
 
 ### **v1.0.4.7652** 
 -> Update reference package version. <br />
--> Fix some warnings.
+-> Fix some warnings.<br />
 
 ### **v1.0.5.5444** 
--> Update reference package version, fixing CVE (`CVE-2024-43485`).
+-> Update reference package version, fixing CVE (`CVE-2024-43485`).<br />
+
+### **v1.1.0.0** 
+-> Add new/edit data type extensions.<br />
+-> Upgrade ref package version.<br />
+-> Add and adjust existing message store and HTTP code rfc link.<br />
+-> Add new ProblemDetails result model. Extend existing Microsoft.AspNetCore.Mvc.ProblemDetails.<br />
+-> Add cast helper from `Result` > `ProblemDetails`. Add API `ObjectResult` extension methods.<br />